Use absolute path to load the YAML configuration
Ocasionally one might run the vagrant command with a different cwd then
the Vagrantfile is located.

Vagrant goes up a directory tree to find the Vagrantfile, but when we
open the YAML files the path is relative to the cwd.

This change makes sure that we get the right path to load the YAML
configuration.