Add tasks to standardize control repo workflow
During Puppet development, a frequent task is to work with a control
repository. Most of the time and specially when developing for various
operating systems, starting a new environment and make it available to
the target VM often require a series of steps.

The task bellow will ease the process:

rake controlrepo:setup CONTROL_REPO_URL=git://server/control-repo.git

Other task are available to complement:
* controlrepo:create_environment
* controlrepo:deploy_environment
* controlrepo:make_environments
* controlrepo:prepare